Weather data coming from API has been integrated into the collectionview structure in a harmonious way
    - Fix JSON decoding error caused by Type mismatch in Forecast structure.

+++ ForecastApp/Model/Forecast.swift
@@ -17,9 +17,9 @@ struct WeatherModel: Codable {
         let description: String
     }
     struct Main: Codable {
-        let temp: Int32
-        let temp_min: Int32
-        let temp_max: Int32
+        let temp: Double
+        let temp_min: Double
+        let temp_max: Double
     }
     
     static let mockWeatherData: [WeatherModel] = [
